SmoothNumbers

Documentation for SmoothNumbers.

SmoothNumbers.pratt

SmoothNumbers.pratt

SmoothNumbers.pratt

SmoothNumbers.smooth

SmoothNumbers.smooth

SmoothNumbers.smooth

SmoothNumbers.with_bases

SmoothNumbers.with_bases

SmoothNumbers.pratt — Function

Computes 3-smooth numbers.

SmoothNumbers.pratt — Method

pratt(n::Integer)

Computes the first n 3-smooth numbers and returns them as a Vector{Int}.

SmoothNumbers.pratt — Method

pratt(::Type{T}, n::Integer) where T<:Integer

Computes the first n 3-smooth numbers and returns them as a Vector{T}.

SmoothNumbers.smooth — Function

Computes smooth numbers.

SmoothNumbers.smooth — Method

smooth(k::Integer, n::Integer)

Computes the first n k-smooth numbers and returns them as a Vector{typeof(k)}.

SmoothNumbers.smooth — Method

smooth(::Type{T}, k::Integer, n::Integer) where T<:Integer

Computes the first n k-smooth numbers and returns them as a Vector{T}.

SmoothNumbers.with_bases — Function

Computes numbers that can be written as a product of powers of a given set of bases.

SmoothNumbers.with_bases — Method

with_bases(bases::AbstractVector{T}, n::Integer) where T<:Integer

Computes the first n numbers that can be written as a product of powers of the bases.

The result is a Vector with the same element type as bases.